Refactored ImageUpload into own component
