+++ src/components/Forms/ImageUpload/style.module.less
@@ -1,0 +1,82 @@
+@import '../../style/vars.less';
+
+.avatarImage {
+  height: 5rem;
+  width: 5rem;
+  display: block;
+  margin-top: 1rem;
+  transition: transform 0.3s;
+  pointer-events: all;
+
+  @media (max-width: @breakPointM) {
+    height: 4rem;
+    width: 4rem;
+  }
+}
+
+.avatarUploadButton {
+  visibility: hidden;
+  position: absolute;
+}
+
+.avatarImageContainer {
+  position: relative;
+  cursor: pointer;
+  pointer-events: none;
+
+  &:hover {
+    .avatarImage {
+      transform: rotate(-15deg);
+    }
+  }
+}
+
+.avatarImageLabel {
+  position: absolute;
+  top: calc(100% + 6.5rem);
+  left: calc(100% + 4.5rem);
+  width: 15rem;
+  font-size: @fontSizeS;
+  pointer-events: all;
+
+  @media (max-width: @breakPointM) {
+    transform: rotate(70deg);
+    transform-origin: 0 0;
+    top: 100%;
+    left: 2rem;
+    font-size: @fontSizeXS;
+  }
+
+  &:hover {
+    color: @blueBright;
+  }
+
+  &:before {
+    content: '';
+    position: absolute;
+    right: 100%;
+    background-image: url('./arrow.svg');
+    background-size: 100% 100%;
+    bottom: 0.8rem;
+    width: 3rem;
+    height: 2rem;
+
+    @media (max-width: @breakPointM) {
+      left: -2.1rem;
+      bottom: -0.8rem;
+      right: auto;
+      transform: rotate(-20deg);
+    }
+  }
+}
+
+.submitButton {
+  visibility: hidden;
+  position: relative;
+  top: -4rem;
+  left: 7rem;
+}
+
+.submitButtonDirty {
+  visibility: visible;
+}
